moving heath format check
